hedge_fund/requirements.txt
# Data
yfinance>=0.2.31
pandas>=2.0.0
numpy>=1.24.0
# pandas-ta removed — indicators computed directly with pandas/numpy

# PDF
pdfplumber>=0.10.0

# Visualization & Reporting
matplotlib>=3.7.0
seaborn>=0.12.0
jinja2>=3.1.0

# Brokers
alpaca-py>=0.21.0

# Config
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0

# Utilities
pyarrow>=14.0.0

# Optional: Interactive Brokers
# ib_insync>=0.9.86
